Refactored `navigation.py` and implemented PR review feedback
